Add OAuth library
The OAuth library contains functions that handle generating authorization URLs,
adding tokens to a user, getting the token of a user, and refreshing a token
when a user's token has expired. The nice part is that you don't have worry
about tracking whether or not a token has expired.

The most helpful function is Auth::getToken(), for when you need to add their
OAuth token to an Authorization header.